# Increase efficiency of preassembly combine_related by more restrictive statement grouping