--- a/indra/preassembler/__init__.py
+++ b/indra/preassembler/__init__.py
@@ -392,7 +392,6 @@ class Preassembler(object):
         logger.debug("Total comparisons: %s" % total_comps)
         if group_sizes:
             logger.debug("Max group size: %s" % np.max(group_sizes))
-            logger.debug("Largest group: %s" % str(largest_group))
             logger.debug("(%.1f %% of all comparisons)" %
                   (100 * ((np.max(group_sizes) ** 2) / float(total_comps))))